requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14
sendgrid>=6.10
django-cors-headers>=4.3
python-dotenv>=1.0
